Deprecate RecurrencePatternSerializer
